Setting up backend integration for Profile
